+++ Changelog.md
@@ -45,6 +45,7 @@
 * Add parsing `ParagraphProperties#justification`
 * Add parsing `TableRowProperties#table_header` (fix #264)
 * Add parsing `XlsxColumnProperties#custom_width`, `XlsxColumnProperties#best_fit`
+* Add parsing `DocxParagraphRun#object` and `Object#ole_object`
 * New way to parse default RunProperties and ParagraphProperties. Old way is still there
 
 ### Fixes